Make vortex initial condition importable
