+++ grudge/models/euler.py
@@ -40,6 +40,8 @@ from grudge.dof_desc import DOFDesc, DISCR_TAG_BASE, as_dofdesc
 from grudge.models import HyperbolicOperator
 from grudge.trace_pair import TracePair
 
+from pytools.obj_array import make_obj_array
+
 import grudge.op as op
 
 
@@ -63,6 +65,39 @@ class EulerContainer:
     @property
     def dim(self):
         return len(self.momentum)
+
+# }}}
+
+
+# {{{ Predefined initial conditions for the Euler model
+
+def vortex_initial_condition(x_vec, t=0):
+    """Initial condition adapted from Section 2 (equation 2) of:
+
+    - K. Mattsson, M. Sv\"{a}rd, M. Carpenter, and J. Nordstr\"{o}m (2006).
+    High-order accurate computations for unsteady aerodynamics.
+    [DOI](https://doi.org/10.1016/j.compfluid.2006.02.004).
+    """
+    mach = 0.5    # Mach number
+    _x0 = 5
+    epsilon = 1   # vortex strength
+    gamma = 1.4
+    x, y = x_vec
+    actx = x.array_context
+
+    fxyt = 1 - (((x - _x0) - t)**2 + y**2)
+    expterm = actx.np.exp(fxyt/2)
+
+    u = 1 - (epsilon*y/(2*np.pi)) * expterm
+    v = ((epsilon*(x - _x0) - t)/(2*np.pi)) * expterm
+
+    velocity = make_obj_array([u, v])
+    mass = (
+        1 - ((epsilon**2 * (gamma - 1) * mach**2)/(8*np.pi**2)) * actx.np.exp(fxyt)
+    ) ** (1 / (gamma - 1))
+    p = (mass ** gamma)/(gamma * mach**2)
+
+    return primitive_to_conservative_vars((mass, velocity, p), gamma=gamma)
 
 # }}}
